requirements.txt
rembg
numpy
pillow
onnxruntime
boto3
